chore(api): align org-read survey helpers
