+++ apps/web/app/api/v1/management/surveys/lib/utils.test.ts
@@ -1,4 +1,5 @@
 import { afterEach, describe, expect, test, vi } from "vitest";
+import type { TAuthenticationApiKey } from "@formbricks/types/auth";
 import { TOrganization } from "@formbricks/types/organizations";
 import {
   TSurvey,
@@ -9,7 +10,8 @@ import { responses } from "@/app/lib/api/response";
 import { getIsSpamProtectionEnabled } from "@/modules/ee/license-check/lib/utils";
 import { getSurveyFollowUpsPermission } from "@/modules/survey/follow-ups/lib/utils";
 import { getExternalUrlsPermission } from "@/modules/survey/lib/permission";
-import { checkFeaturePermissions } from "./utils";
+import { getEnvironmentIdsByOrganizationId } from "./environment";
+import { checkFeaturePermissions, getReadableEnvironmentIds } from "./utils";
 
 // Mock dependencies
 vi.mock("@/app/lib/api/response", () => ({
@@ -32,6 +34,10 @@ vi.mock("@/modules/survey/lib/permission", () => ({
 
 vi.mock("@/lib/survey/utils", () => ({
   getElementsFromBlocks: vi.fn((blocks: any[]) => blocks.flatMap((block: any) => block.elements)),
+}));
+
+vi.mock("./environment", () => ({
+  getEnvironmentIdsByOrganizationId: vi.fn(),
 }));
 
 const mockOrganization: TOrganization = {
@@ -107,6 +113,109 @@ const baseSurveyData: TSurveyCreateInputWithEnvironmentId = {
   blocks: [],
   followUps: [],
 };
+
+const baseAuthentication = {
+  type: "apiKey" as const,
+  apiKeyId: "api-key-id",
+  organizationId: "org-id",
+  organizationAccess: {
+    accessControl: {
+      read: false,
+      write: false,
+    },
+  },
+  environmentPermissions: [],
+};
+
+const environmentPermission = (
+  environmentId: string,
+  permission: "read" | "write" | "manage"
+): TAuthenticationApiKey["environmentPermissions"][number] => ({
+  environmentId,
+  permission,
+  environmentType: "development",
+  projectId: `project-${environmentId}`,
+  projectName: `Project ${environmentId}`,
+});
+
+describe("getReadableEnvironmentIds", () => {
+  afterEach(() => {
+    vi.clearAllMocks();
+  });
+
+  test("returns all organization environments when API key has organization read access", async () => {
+    vi.mocked(getEnvironmentIdsByOrganizationId).mockResolvedValue(["env-1", "env-2"]);
+
+    const result = await getReadableEnvironmentIds({
+      ...baseAuthentication,
+      organizationAccess: {
+        accessControl: {
+          read: true,
+          write: false,
+        },
+      },
+    });
+
+    expect(result).toEqual(["env-1", "env-2"]);
+    expect(getEnvironmentIdsByOrganizationId).toHaveBeenCalledWith("org-id");
+  });
+
+  test("returns an empty list when an organization-read API key belongs to an organization without environments", async () => {
+    vi.mocked(getEnvironmentIdsByOrganizationId).mockResolvedValue([]);
+
+    const result = await getReadableEnvironmentIds({
+      ...baseAuthentication,
+      organizationAccess: {
+        accessControl: {
+          read: true,
+          write: false,
+        },
+      },
+    });
+
+    expect(result).toEqual([]);
+    expect(getEnvironmentIdsByOrganizationId).toHaveBeenCalledWith("org-id");
+  });
+
+  test("returns all organization environments when API key has organization write access", async () => {
+    vi.mocked(getEnvironmentIdsByOrganizationId).mockResolvedValue(["env-1"]);
+
+    const result = await getReadableEnvironmentIds({
+      ...baseAuthentication,
+      organizationAccess: {
+        accessControl: {
+          read: false,
+          write: true,
+        },
+      },
+    });
+
+    expect(result).toEqual(["env-1"]);
+    expect(getEnvironmentIdsByOrganizationId).toHaveBeenCalledWith("org-id");
+  });
+
+  test("returns de-duplicated environment permissions that allow GET without organization access", async () => {
+    const result = await getReadableEnvironmentIds({
+      ...baseAuthentication,
+      environmentPermissions: [
+        environmentPermission("env-1", "read"),
+        environmentPermission("env-2", "write"),
+        environmentPermission("env-3", "manage"),
+        environmentPermission("env-1", "read"),
+      ],
+    });
+
+    expect(result).toEqual(["env-1", "env-2", "env-3"]);
+    expect(getEnvironmentIdsByOrganizationId).not.toHaveBeenCalled();
+  });
+
+  test("returns null when the API key has no readable access", async () => {
+    const result = await getReadableEnvironmentIds(baseAuthentication);
+
+    expect(result).toBeNull();
+    expect(getEnvironmentIdsByOrganizationId).not.toHaveBeenCalled();
+  });
+});
 
 describe("checkFeaturePermissions", () => {
   vi.mocked(getExternalUrlsPermission).mockResolvedValue(true);
