+++ apps/web/app/api/v1/management/surveys/lib/utils.ts
@@ -1,10 +1,32 @@
+import { OrganizationAccessType } from "@formbricks/types/api-key";
+import type { TAuthenticationApiKey } from "@formbricks/types/auth";
 import { TOrganization } from "@formbricks/types/organizations";
 import { TSurvey, TSurveyCreateInputWithEnvironmentId } from "@formbricks/types/surveys/types";
 import { responses } from "@/app/lib/api/response";
 import { getElementsFromBlocks } from "@/lib/survey/utils";
 import { getIsSpamProtectionEnabled } from "@/modules/ee/license-check/lib/utils";
+import { hasOrganizationAccess, hasPermission } from "@/modules/organization/settings/api-keys/lib/utils";
 import { getSurveyFollowUpsPermission } from "@/modules/survey/follow-ups/lib/utils";
 import { getExternalUrlsPermission } from "@/modules/survey/lib/permission";
+import { getEnvironmentIdsByOrganizationId } from "./environment";
+
+export const getReadableEnvironmentIds = async (
+  authentication: TAuthenticationApiKey
+): Promise<string[] | null> => {
+  if (hasOrganizationAccess(authentication, OrganizationAccessType.Read)) {
+    return getEnvironmentIdsByOrganizationId(authentication.organizationId);
+  }
+
+  const environmentIds = authentication.environmentPermissions
+    .filter((permission) =>
+      hasPermission(authentication.environmentPermissions, permission.environmentId, "GET")
+    )
+    .map((permission) => permission.environmentId);
+
+  const readableEnvironmentIds = Array.from(new Set(environmentIds));
+
+  return readableEnvironmentIds.length > 0 ? readableEnvironmentIds : null;
+};
 
 export const checkFeaturePermissions = async (
   surveyData: TSurveyCreateInputWithEnvironmentId,
